Separate networking tools that may be used by modules outside of networking so changes to networking-only utilities don't trigger AWS integration tests
